+++ src/client/interfaces/ifield-run.iface.js
@@ -342,7 +342,7 @@ export const IFieldRun = DeclareMixin(( superclass ) => class extends superclass
         if( !this.#showStatus ){
             let display = this.iRunChecker().confDisplayStatus();
             //console.debug( 'iRunShowStatus checker display', this.name(), display );
-            const overridable = Forms.configure().checkStatusOverridable;
+            const overridable = Forms.configure().showStatusOverridable;
             //console.debug( 'iRunShowStatus overridable', this.name(), overridable );
             if( overridable ){
                 const status = this.iSpecStatus();